+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/RequestFactoryBase.java
@@ -4,7 +4,11 @@ import java.net.http.HttpRequest;
 import java.net.http.HttpRequest.Builder;
 import java.util.Arrays;
 import java.util.Map;
+import java.util.Optional;
+import java.util.Properties;
+import java.util.stream.Collectors;
 
+import lombok.extern.slf4j.Slf4j;
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.util.FlinkRuntimeException;
@@ -14,10 +18,12 @@ import com.getindata.connectors.http.LookupQueryCreator;
 import com.getindata.connectors.http.internal.HeaderPreprocessor;
 import com.getindata.connectors.http.internal.config.HttpConnectorConfigConstants;
 import com.getindata.connectors.http.internal.utils.HttpHeaderUtils;
+import static com.getindata.connectors.http.internal.table.lookup.HttpLookupConnectorOptions.SOURCE_LOOKUP_OIDC_AUTH_TOKEN_ENDPOINT_URL;
 
 /**
  * Base class for {@link HttpRequest} factories.
  */
+@Slf4j
 public abstract class RequestFactoryBase implements HttpRequestFactory {
 
     public static final String DEFAULT_REQUEST_TIMEOUT_SECONDS = "30";
@@ -35,6 +41,7 @@ public abstract class RequestFactoryBase implements HttpRequestFactory {
      * HTTP headers that should be used for {@link HttpRequest} created by factory.
      */
     private final String[] headersAndValues;
+    private final HttpLookupConfig options;
 
     public RequestFactoryBase(
             LookupQueryCreator lookupQueryCreator,
@@ -43,6 +50,21 @@ public abstract class RequestFactoryBase implements HttpRequestFactory {
 
         this.baseUrl = options.getUrl();
         this.lookupQueryCreator = lookupQueryCreator;
+        this.options = options;
+
+        Properties properties = options.getProperties();
+        /*
+         * For OIDC, the preprocessor will fully specify the Authentication header value,
+         * as a bearer token. But the preprocessors only amend existing headers, so in this case
+         * if there is no existing authorization header then we add a dummy one to the properties,
+         * so the preprocessor will be driven and will provide the value.
+         */
+        Optional<String> oidcAuthURL = options.getReadableConfig()
+                .getOptional(SOURCE_LOOKUP_OIDC_AUTH_TOKEN_ENDPOINT_URL);
+        if (oidcAuthURL.isPresent()) {
+            properties.put(HttpConnectorConfigConstants.LOOKUP_SOURCE_HEADER_PREFIX
+                    + HttpHeaderUtils.AUTHORIZATION, "Dummy");
+        }
 
         var headerMap = HttpHeaderUtils
             .prepareHeaderMap(
@@ -52,6 +74,11 @@ public abstract class RequestFactoryBase implements HttpRequestFactory {
             );
 
         this.headersAndValues = HttpHeaderUtils.toHeaderAndValueArray(headerMap);
+
+        log.debug("RequestFactoryBase headersAndValues: " +
+                Arrays.stream(headersAndValues)
+                        .map(Object::toString)
+                        .collect(Collectors.joining(",")));
         this.httpRequestTimeOutSeconds = Integer.parseInt(
             options.getProperties().getProperty(
                 HttpConnectorConfigConstants.LOOKUP_HTTP_TIMEOUT_SECONDS,
@@ -67,7 +94,6 @@ public abstract class RequestFactoryBase implements HttpRequestFactory {
         getLogger().debug("Created Http lookup query: " + lookupQueryInfo);
 
         Builder requestBuilder = setUpRequestMethod(lookupQueryInfo);
-
         if (headersAndValues.length != 0) {
             requestBuilder.headers(headersAndValues);
         }